Ensure binary mode for output on Windows
On Windows the EOL convention is '\r\n', which means that when we output
a newline '\n', because we want a line feed in the output, it gets turned
into a carriage return line feed.  This isn't good for ESC sequences, so
adjust the mode of stdout to be binary on Windows, which prevents this.